requirements.txt
Django==6.0.1
django-tailwind==4.4.2
asgiref==3.11.0
sqlparse==0.5.5
python-dotenv==1.0.1
PyNaCl
base58
